improve test coverage (#868)
* Adding "pragma: no cover" to ImportError stanzas

* Improving coverage in utils.py

* Added a test for ufuncs/s3crypt.py

* Commenting out dead code in schema.py

* In schema.py yanked a stray pdb

* Commenting out `access` subcommand in commands.py for now.

* Adding coverage for tag filter

* More coverage for emr.py error conditions

* Added emr coverage.  Combined ec2 tests

* Adding coverage for ebs.py

* Fixed ebs skip-ami-snapshots filter validate() method and added another test for it.

* Revert "Fixed ebs skip-ami-snapshots filter validate() method and added another test for it."

This reverts commit b08718cdb4ea63d933d664b3df9ed5190a68a39b.

* Changing how I write the failure test for EBS (because it is breaking on travis CI, but not locally)

* Fix for another EBS test that fails only on travis CI

* Fixed ebs skip-ami-snapshots filter validate() method and added another test for it.

* Fixed ebs validation test.  Added VPC validation test

* In actions.py, fix 'NotImplemented' -> 'NotImplementedError'.  Added tests for actions.py coverage

* Fixing autotag test

* Adding coverage for output.py

+++ tests/common.py
@@ -57,16 +57,17 @@ class BaseTest(PillTest):
 
     def load_policy(
             self, data, config=None, session_factory=None,
-            validate=C7N_VALIDATE, cache=False):
+            validate=C7N_VALIDATE, output_dir=None, cache=False):
         if validate:
             errors = schema_validate({'policies': [data]}, C7N_SCHEMA)
             if errors:
                 raise errors[0]
 
         config = config or {}
-        temp_dir = tempfile.mkdtemp()
-        self.addCleanup(shutil.rmtree, temp_dir)
-        config['output_dir'] = temp_dir
+        if not output_dir:
+            temp_dir = tempfile.mkdtemp()
+            self.addCleanup(shutil.rmtree, temp_dir)
+            config['output_dir'] = temp_dir
         if cache:
             config['cache'] = os.path.join(temp_dir, 'c7n.cache')
             config['cache_period'] = 300
